+++ props/core/BUILD.bazel
@@ -241,9 +241,11 @@ docker_py_test(
         "//props:conftest",
         "//props/core:oci_utils",
         "//props/core/models:examples",
+        "//props/critic/testing:mocks",
         "//props/critic_dev:loop",
         "//props/critic_dev:shared",
         "//props/critic_dev/optimize:orchestration_fixtures",
+        "//props/critic_dev/testing:mocks",
         "//props/db:database",
         "//props/db:models",
         "//props/testing",
