+++ props/core/test_agent_pkg_e2e.py
@@ -36,6 +36,7 @@ from props.core.eval_api_models import GradingStatusResponse, RunCriticResponse
 from props.core.ids import SnapshotSlug
 from props.core.models.examples import ExampleKind, WholeSnapshotExample
 from props.core.oci_utils import BUILTIN_TAG
+from props.critic.testing.mocks import CriticMock
 from props.critic_dev.loop import RunCriticToolArgs, WaitUntilGradedToolArgs
 from props.critic_dev.optimize.orchestration_fixtures import (
     ORCHESTRATION_CRITIC_MODEL,
@@ -44,9 +45,10 @@ from props.critic_dev.optimize.orchestration_fixtures import (
     make_orchestration_grader_mock,
 )
 from props.critic_dev.shared import TargetMetric
+from props.critic_dev.testing.mocks import CriticDevMock
 from props.db.database import Database
 from props.db.models import AgentRun, AgentRunStatus
-from props.testing.mocks import CriticDevMock, CriticMock, get_system_message_text
+from props.testing.mocks import get_system_message_text
 
 logger = logging.getLogger(__name__)
