+++ props/critic/testing/BUILD.bazel
@@ -1,0 +1,16 @@
+"""Shared test fixtures for critic agent."""
+
+load("@rules_python//python:defs.bzl", "py_library")
+
+py_library(
+    name = "mocks",
+    testonly = True,
+    srcs = ["mocks.py"],
+    imports = ["../../.."],
+    visibility = ["//props:__subpackages__"],
+    deps = [
+        "//openai_utils:model",
+        "//props/critic:main",
+        "//props/testing",
+    ],
+)
