+++ props/critic_dev/optimize/BUILD.bazel
@@ -82,8 +82,10 @@ docker_py_test(
         "//props/core:eval_api_models",
         "//props/core:oci_utils",
         "//props/core/models:examples",
+        "//props/critic/testing:mocks",
         "//props/critic_dev:loop",
         "//props/critic_dev:shared",
+        "//props/critic_dev/testing:mocks",
         "//props/db:config",
         "//props/db:database",
         "//props/db:examples",
